+++ activerecord/lib/active_record/connection_adapters/abstract_mysql_adapter.rb
@@ -526,22 +526,15 @@ module ActiveRecord
         index.using == :btree || super
       end
 
-      def insert_fixtures(*)
-        without_sql_mode("NO_AUTO_VALUE_ON_ZERO") { super }
-      end
-
       def insert_fixtures_set(fixture_set, tables_to_delete = [])
         iterate_over_results = -> { while raw_connection.next_result; end; }
 
         with_multi_statements do
-          without_sql_mode("NO_AUTO_VALUE_ON_ZERO") do
-            super(fixture_set, tables_to_delete, &iterate_over_results)
-          end
+          super(fixture_set, tables_to_delete, &iterate_over_results)
         end
       end
 
       private
-
         def combine_multi_statements(total_sql)
           total_sql.each_with_object([]) do |sql, total_sql_chunks|
             previous_packet = total_sql_chunks.last
@@ -578,19 +571,6 @@ module ActiveRecord
         ensure
           @config[:flags] = previous_flags
           reconnect!
-        end
-
-        def without_sql_mode(mode)
-          result = execute("SELECT @@SESSION.sql_mode")
-          current_mode = result.first[0]
-          return yield unless current_mode.include?(mode)
-
-          sql_mode = "REPLACE(@@sql_mode, '#{mode}', '')"
-          execute("SET @@SESSION.sql_mode = #{sql_mode}")
-          yield
-        ensure
-          sql_mode = "CONCAT(@@sql_mode, ',#{mode}')"
-          execute("SET @@SESSION.sql_mode = #{sql_mode}")
         end
 
         def initialize_type_map(m = type_map)