+++ activerecord/test/cases/adapters/postgresql/change_schema_test.rb
@@ -33,7 +33,7 @@ module ActiveRecord
         connection.change_column :strings, :somedate, :timestamp, array: true, cast_as: :timestamp
         column = connection.columns(:strings).find { |c| c.name == "somedate" }
         assert_equal :datetime, column.type
-        assert column.array?
+        assert_predicate column, :array?
       end
     end
   end