+++ activerecord/test/cases/persistence_test.rb
@@ -217,7 +217,7 @@ class PersistenceTest < ActiveRecord::TestCase
   def test_destroy_all
     conditions = "author_name = 'Mary'"
     topics_by_mary = Topic.all.merge!(where: conditions, order: "id").to_a
-    assert ! topics_by_mary.empty?
+    assert_not_empty topics_by_mary
 
     assert_difference("Topic.count", -topics_by_mary.size) do
       destroyed = Topic.where(conditions).destroy_all.sort_by(&:id)
@@ -253,7 +253,7 @@ class PersistenceTest < ActiveRecord::TestCase
 
   def test_becomes_includes_errors
     company = Company.new(name: nil)
-    assert !company.valid?
+    assert_not_predicate company, :valid?
     original_errors = company.errors
     client = company.becomes(Client)
     assert_equal original_errors.keys, client.errors.keys
@@ -370,8 +370,8 @@ class PersistenceTest < ActiveRecord::TestCase
     developer.destroy
     new_developer = developer.dup
     new_developer.save
-    assert new_developer.persisted?
-    assert_not new_developer.destroyed?
+    assert_predicate new_developer, :persisted?
+    assert_not_predicate new_developer, :destroyed?
   end
 
   def test_create_many
@@ -387,7 +387,7 @@ class PersistenceTest < ActiveRecord::TestCase
     )
     topic = topic.dup # reset @new_record
     assert_nothing_raised { topic.save }
-    assert topic.persisted?
+    assert_predicate topic, :persisted?
     assert_equal "Another New Topic", topic.reload.title
   end
 
@@ -435,7 +435,7 @@ class PersistenceTest < ActiveRecord::TestCase
     topic_reloaded = Topic.instantiate(topic.attributes.merge("does_not_exist" => "test"))
     topic_reloaded.title = "A New Topic"
     assert_nothing_raised { topic_reloaded.save }
-    assert topic_reloaded.persisted?
+    assert_predicate topic_reloaded, :persisted?
     assert_equal "A New Topic", topic_reloaded.reload.title
   end
 
@@ -617,63 +617,63 @@ class PersistenceTest < ActiveRecord::TestCase
   def test_delete_new_record
     client = Client.new(name: "37signals")
     client.delete
-    assert client.frozen?
+    assert_predicate client, :frozen?
 
     assert_not client.save
     assert_raise(ActiveRecord::RecordNotSaved) { client.save! }
 
-    assert client.frozen?
+    assert_predicate client, :frozen?
     assert_raise(RuntimeError) { client.name = "something else" }
   end
 
   def test_delete_record_with_associations
     client = Client.find(3)
     client.delete
-    assert client.frozen?
+    assert_predicate client, :frozen?
     assert_kind_of Firm, client.firm
 
     assert_not client.save
     assert_raise(ActiveRecord::RecordNotSaved) { client.save! }
 
-    assert client.frozen?
+    assert_predicate client, :frozen?
     assert_raise(RuntimeError) { client.name = "something else" }
   end
 
   def test_destroy_new_record
     client = Client.new(name: "37signals")
     client.destroy
-    assert client.frozen?
+    assert_predicate client, :frozen?
 
     assert_not client.save
     assert_raise(ActiveRecord::RecordNotSaved) { client.save! }
 
-    assert client.frozen?
+    assert_predicate client, :frozen?
     assert_raise(RuntimeError) { client.name = "something else" }
   end
 
   def test_destroy_record_with_associations
     client = Client.find(3)
     client.destroy
-    assert client.frozen?
+    assert_predicate client, :frozen?
     assert_kind_of Firm, client.firm
 
     assert_not client.save
     assert_raise(ActiveRecord::RecordNotSaved) { client.save! }
 
-    assert client.frozen?
+    assert_predicate client, :frozen?
     assert_raise(RuntimeError) { client.name = "something else" }
   end
 
   def test_update_attribute
-    assert !Topic.find(1).approved?
+    assert_not_predicate Topic.find(1), :approved?
     Topic.find(1).update_attribute("approved", true)
-    assert Topic.find(1).approved?
+    assert_predicate Topic.find(1), :approved?
 
     Topic.find(1).update_attribute(:approved, false)
-    assert !Topic.find(1).approved?
+    assert_not_predicate Topic.find(1), :approved?
 
     Topic.find(1).update_attribute(:change_approved_before_save, true)
-    assert Topic.find(1).approved?
+    assert_predicate Topic.find(1), :approved?
   end
 
   def test_update_attribute_for_readonly_attribute
@@ -710,14 +710,14 @@ class PersistenceTest < ActiveRecord::TestCase
   def test_update_column
     topic = Topic.find(1)
     topic.update_column("approved", true)
-    assert topic.approved?
+    assert_predicate topic, :approved?
     topic.reload
-    assert topic.approved?
+    assert_predicate topic, :approved?
 
     topic.update_column(:approved, false)
-    assert !topic.approved?
+    assert_not_predicate topic, :approved?
     topic.reload
-    assert !topic.approved?
+    assert_not_predicate topic, :approved?
   end
 
   def test_update_column_should_not_use_setter_method
@@ -804,10 +804,10 @@ class PersistenceTest < ActiveRecord::TestCase
   def test_update_columns
     topic = Topic.find(1)
     topic.update_columns("approved" => true, title: "Sebastian Topic")
-    assert topic.approved?
+    assert_predicate topic, :approved?
     assert_equal "Sebastian Topic", topic.title
     topic.reload
-    assert topic.approved?
+    assert_predicate topic, :approved?
     assert_equal "Sebastian Topic", topic.title
   end
 
@@ -915,33 +915,33 @@ class PersistenceTest < ActiveRecord::TestCase
 
   def test_update
     topic = Topic.find(1)
-    assert !topic.approved?
+    assert_not_predicate topic, :approved?
     assert_equal "The First Topic", topic.title
 
     topic.update("approved" => true, "title" => "The First Topic Updated")
     topic.reload
-    assert topic.approved?
+    assert_predicate topic, :approved?
     assert_equal "The First Topic Updated", topic.title
 
     topic.update(approved: false, title: "The First Topic")
     topic.reload
-    assert !topic.approved?
+    assert_not_predicate topic, :approved?
     assert_equal "The First Topic", topic.title
   end
 
   def test_update_attributes
     topic = Topic.find(1)
-    assert !topic.approved?
+    assert_not_predicate topic, :approved?
     assert_equal "The First Topic", topic.title
 
     topic.update_attributes("approved" => true, "title" => "The First Topic Updated")
     topic.reload
-    assert topic.approved?
+    assert_predicate topic, :approved?
     assert_equal "The First Topic Updated", topic.title
 
     topic.update_attributes(approved: false, title: "The First Topic")
     topic.reload
-    assert !topic.approved?
+    assert_not_predicate topic, :approved?
     assert_equal "The First Topic", topic.title
 
     error = assert_raise(ActiveRecord::RecordNotUnique, ActiveRecord::StatementInvalid) do
@@ -1042,7 +1042,7 @@ class PersistenceTest < ActiveRecord::TestCase
     Topic.find(1).replies << should_be_destroyed_reply
 
     topic = Topic.destroy(1)
-    assert topic.destroyed?
+    assert_predicate topic, :destroyed?
 
     assert_raise(ActiveRecord::RecordNotFound) { Topic.find(1) }
     assert_raise(ActiveRecord::RecordNotFound) { Reply.find(should_be_destroyed_reply.id) }
@@ -1122,13 +1122,13 @@ class PersistenceTest < ActiveRecord::TestCase
   def test_find_via_reload
     post = Post.new
 
-    assert post.new_record?
+    assert_predicate post, :new_record?
 
     post.id = 1
     post.reload
 
     assert_equal "Welcome to the weblog", post.title
-    assert_not post.new_record?
+    assert_not_predicate post, :new_record?
   end
 
   def test_reload_via_querycache